qt: fix incorrect param type
which was causing a whole bunch of "'< 0' is always false" warnings.

review of the original fix ([1]) indicated that in fact the parameter was
wrong rather than the checks within the function being out of step with
the final design of the function interface, since apparently callers
are in fact passing in a signed index, where a negative represents an
invalid index which the function is expected to handle appropriately.

[1]: https://mailman.videolan.org/pipermail/vlc-devel/2020-September/137711.html

+++ modules/gui/qt/medialibrary/mlalbummodel.cpp
@@ -38,7 +38,7 @@ QVariant MLAlbumModel::data(const QModelIndex &index, int role) const
     if (!index.isValid() || index.row() < 0)
         return QVariant();
 
-    const MLAlbum* ml_item = item(static_cast<unsigned int>(index.row()));
+    const MLAlbum* ml_item = item(index.row());
     if ( ml_item == NULL )
         return QVariant();
